pyproject.toml
[tool.poetry]
name = "langchain practice"
version = "0.1.0"
description = ""
authors = ["Caephas <arinzeobidiegwu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.3.15"
langchain-community = "^0.3.15"
langchain-google-firestore = "^0.3.0"
firestore = "^0.0.8"
chromadb = "^0.5.0"
tiktoken = "^0.7.0"
sentence-transformers = "^3.0.0"
bs4 = "^0.0.2"
firecrawl-py = "^0.0.13"
langchainhub = "^0.1.18"
gpt4all = "0.3.3"
llama-cpp-python = "^0.3.6"
wikipedia = "^1.4.0"
tavily-python = "^0.3.3"
ipykernel = "^6.29.5"
notebook = "^7.3.2"
pypdf = "^5.1.0"
langchain-google-genai = "2.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
